stop playground pipeline when leaving the playground ui page
